put bracket logic in a model
